fix: search_files
Check if files is directory

+++ helpers/filemanager.py
@@ -54,11 +54,12 @@ def copy_files(files):
     filenames = []
     prefix = socket.gethostname() + '_' + getpass.getuser() + '_'
     for file in files:
-        if os.path.normpath("PyWeasel") not in file and os.stat(file).st_size >= 1:
-            ext = os.path.splitext(file)[1]
-            filename = slugify(os.path.splitext(file)[0]) + ext
-            shutil.copy(file, os.path.join(sfiles_dir_exist(), filename))
-            filenames.append((file, filename))
+        if os.path.isfile(file):
+            if os.path.normpath("PyWeasel") not in file and os.stat(file).st_size >= 1:
+                ext = os.path.splitext(file)[1]
+                filename = slugify(os.path.splitext(file)[0]) + ext
+                shutil.copy(file, os.path.join(sfiles_dir_exist(), filename))
+                filenames.append((file, filename))
     return filenames
